src/data/scrapers/stats_scraper.py
"""Stats scraper for team and player statistics"""

from typing import Optional, Dict, Any
import requests
from bs4 import BeautifulSoup

from src.data.models import TeamStats
from src.utils.logging import get_logger

logger = get_logger("scrapers.stats")


class StatsScraper:
    """Scraper for team statistics"""
    
    def __init__(self):
        """Initialize stats scraper"""
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        }
    
    def get_team_stats(self, team: str) -> Optional[TeamStats]:
        """Get team statistics"""
        logger.info(f"Fetching stats for {team}")
        
        try:
            # Placeholder for actual stats scraping
            # Could scrape from Sports Reference, KenPom, etc.
            return self._get_mock_stats(team)
        except Exception as e:
            logger.error(f"Error fetching stats for {team}: {e}")
            return self._get_mock_stats(team)
    
    def _get_mock_stats(self, team: str) -> TeamStats:
        """Get mock team stats for testing"""
        return TeamStats(
            team=team,
            wins=15,
            losses=5,
            points_per_game=75.5,
            points_allowed_per_game=68.2,
            offensive_rating=110.5,
            defensive_rating=98.3,
            pace=68.5
        )
    
    def get_injury_report(self, team: str) -> list:
        """Get injury report for a team"""
        logger.info(f"Fetching injury report for {team}")
        
        # Placeholder for actual injury scraping
        # Would scrape from ESPN, team websites, etc.
        return []

