requirements.txt
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pyyaml>=6.0
lxml>=4.9.0

